0528-random-pick-with-weight Medium .cpp Array
Solved by: vaibhav_jpd 
LinkedIn: https://www.linkedin.com/in/vaibhav-raj-singh-/

LeetCode: leetcode.com/u/rosvert/

Portfolio:
https://www.vaibhavrajsingh.in/

#vaibhav-raj-singh
#jiit
#Jaypee

Student @Jaypee Institute of Information Technology

+++ C++/Medium/0528-random-pick-with-weight/0528-random-pick-with-weight.cpp
@@ -1,9 +1,9 @@
 class Solution {
 private:
-    vector<long long int> nums;
+    vector<int> nums;
 public:
     Solution(vector<int>& w) {
-        long long int prev = 0;
+        int prev = 0;
         for(auto it : w){
             prev += it;
             nums.push_back(prev);
